refactor breadcrumb control
