+++ src/routes/browse/[...path]/+page.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
-	import { createElementId } from '$lib/utils.js';
 	import {
-	Button,
+		Button,
 		Col,
 		Collapse,
 		Container,
@@ -22,6 +21,7 @@
 	import ThumbnailCard from './ThumbnailCard.svelte';
 	import { navigating, page } from '$app/stores';
 	import { goto } from '$app/navigation';
+	import Breadcrumb from '$lib/Breadcrumb.svelte';
 
 	const { data } = $props();
 
@@ -31,46 +31,11 @@
 		isOpen = event.detail;
 	}
 
-	interface BreadcrumbData {
-		prefix: string;
-		name: string;
-	}
-	function createBreadcrumb(path?: string): BreadcrumbData[] {
-		let output = [];
-		output.push({
-			name: '<root>',
-			prefix: ''
-		});
-
-		if (!path) {
-			return output;
-		}
-
-		let parts = path?.split('/');
-
-		for (let i = 0; i < parts?.length; i++) {
-			const prefix = parts.slice(0, i + 1).join('/');
-
-			output.push({
-				name: parts[i],
-				prefix: prefix
-			});
-		}
-
-		return output;
-	}
-
 	let sortby = $state('name');
 	let order = $state('ascending');
 
-	sortby = $page.url.searchParams.get("sortby") != 'dateTime' ? 'name' : 'dateTime'
-    order = $page.url.searchParams.get("order") != 'descending' ? 'ascending' : 'descending'
-
-	let breadcrumbData: BreadcrumbData[] = $state([]);
-
-	$effect(() => {
-		breadcrumbData = createBreadcrumb(data.path);
-	});
+	sortby = $page.url.searchParams.get('sortby') != 'dateTime' ? 'name' : 'dateTime';
+	order = $page.url.searchParams.get('order') != 'descending' ? 'ascending' : 'descending';
 
 	function moveToHash(node: Element) {
 		const hash = $page.url.hash;
@@ -102,23 +67,9 @@
 			</Nav>
 		</Collapse>
 	</Navbar>
-	<nav aria-label="breadcrumb">
-		<ol class="breadcrumb">
-			{#each breadcrumbData as b, i}
-				{#if i == breadcrumbData.length - 1}
-					<li class="breadcrumb-item active" aria-current="page">
-						{b.name}
-					</li>
-				{:else}
-					<li class="breadcrumb-item">
-						<a href="/browse/{b.prefix}?sortby={sortby}&order={order}#{createElementId(breadcrumbData[i + 1].name)}">
-							{b.name}
-						</a>
-					</li>
-				{/if}
-			{/each}
-		</ol>
-	</nav>
+
+	<Breadcrumb path={data.path} {order} {sortby}></Breadcrumb>
+
 	<InputGroup>
 		<InputGroupText>Order By</InputGroupText>
 		<Input type="select" bind:value={sortby}>
